feat: added synch mechanism for orders at reboot
