+++ src/util/dm_utils/mod.rs
@@ -5,6 +5,7 @@ mod dm_helpers;
 mod notifications_ch_mng;
 mod order_ch_mng;
 
+pub use dm_helpers::seed_admin_chat_last_seen;
 pub use notifications_ch_mng::handle_message_notification;
 pub use order_ch_mng::handle_operation_result;
