+++ src/util/order_utils/mod.rs
@@ -26,7 +26,7 @@ pub use fetch_scheduler::{
 pub use helper::{
     dispute_from_tags, fetch_events_list, get_disputes, get_orders,
     inferred_status_from_trade_action, map_action_to_status, order_from_tags,
-    parse_disputes_events, parse_orders_events,
+    parse_disputes_events, parse_orders_events, validate_range_amount,
 };
 pub use send_new_order::send_new_order;
 pub use take_order::take_order;
